+++ demos/sac.py
@@ -3,7 +3,8 @@ BasicSAC (from cleanrl)
 启动脚本请用: bash ./benchmarks/sac_run_experiments.py
 查看可用参数: python ./demos/sac.py --help
 单独启动训练:
-python ./demos/sac.py --env.env-type gymnasium --env.env-name Hopper-v4 --total-timesteps 100000 --agent.verbose 2
+python ./demos/sac.py --env.env-type gymnasium --env.env-name Hopper-v4 --agent.num-env-steps 100000 --agent.verbose 2
+python ./demos/sac.py --env.env-type dmc --env.env-name walker-walk --agent.num-env-steps 100000 --agent.verbose 2
 """
 import sys
 from pathlib import Path
@@ -24,44 +25,35 @@ class Args:
     agent: SACConfig
     env: EnvConfig
     logger: LogConfig
-    total_timesteps: int = int(1e6)
     debug: bool = False
 
 if __name__ == '__main__':
+    """ Preprocess """
     args: Args = tyro.cli(Args)
     path_manager.build_path_logs(args.agent, args.env, args.debug)
-    envs = make_envs(args.env)
+    envs, eval_envs = make_envs(args.env)
     logger = get_tensorboard_writer(args.logger, args)
-
-    # 将环境的必要参数保存到agent配置中, 便于模型加载时使用
-    args.agent.num_envs = args.env.env_num
-    args.agent.obs_space = envs.single_observation_space
-    args.agent.act_space = envs.single_action_space
 
     """ Train """
     print("[INFO] Start Training, with args:")
     pprint(args)
-    agent = SAC(args.agent, envs, logger)
-    agent.learn(total_timesteps=args.total_timesteps)
+    agent = SAC(cfg=args.agent, envs=envs, eval_envs=eval_envs, env_cfg=args.env, logger=logger)
+    agent.learn()
     path_ckpt = agent.save()
     del agent
+    envs.close()
     print("[INFO] Finish Training.")
 
-    """ Eval """
+    """ Eval Video """
     print("[INFO] Start Evaluation.")
     agent = SAC.load(path_ckpt, args.agent.device)
-    args.env.env_num = 1
+    args.env.eval_env_num = 1
     args.env.capture_video = True
-    envs = make_envs(args.env)
+    agent.cfg.num_eval_episodes = 1
+    _, eval_envs = make_envs(args.env)
+    agent.eval_envs = eval_envs
+    agent.eval()
 
-    obs = envs.reset()[0]
-    for i in range(int(1e5)):
-        action = agent.predict(obs)
-        obs, rewards, terminations, truncations, infos = envs.step(action)
-        if terminations.any() or truncations.any():
-            envs.step(envs.action_space.sample())  # 再执行一步以保存视频
-            break
-    
     """ Log Video """
     if args.logger.use_swanlab:
         path_videos = path_manager.PATH_LOGS / 'videos'